Clean up stale MetalSharp launch plumbing
